Update patch set 1
Patch Set 1:

> FYI I think this CL doesn't include cl/172081442 / fix for
 > https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=6293

I agree that this doesn't include cl/172081442. But now that I look more closely at it, that just fixes a bug in an internal test that is not in github or AOSP (which is why the change didn't come across).

https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=6293 is a bug with the AOSP version, so cl/172081442 won't fix it. Any idea what the fix for 6293 might be? Maybe this is a new bug?

Patch-set: 1
Reviewer: Gerrit User 1012974 <1012974@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0